// src/client/models/BrewingMethod.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * Describes brewing methods
 */
export enum BrewingMethod {
  ESPRESSO = "Espresso",
  CAPPUCCINO = "Cappuccino",
  LATTE = "Latte",
  AMERICANO = "Americano",
  FILTER = "Filter",
  BIALETTI = "Bialetti",
}